+++ cocos/rendering/index.ts
@@ -29,9 +29,6 @@ export { pipeline };
 export * from './pass-phase';
 export * from './render-types';
 
-export { RenderPipeline } from './render-pipeline';
-export { RenderFlow } from './render-flow';
-export { RenderStage } from './render-stage';
 export { PipelineSceneData } from './pipeline-scene-data';
 
 export { InstancedBuffer } from './instanced-buffer';
