+++ cocos/rendering/legacy/index.jsb.ts
@@ -25,6 +25,10 @@
 declare const nr: any;
 declare const jsb: any;
 
+import { OPEN_HARMONY } from 'internal:constants'
+import { legacyCC } from '../../core/global-exports';
+import { ccenum, CCString } from '../../core';
+
 import { 
     ForwardPipeline as NrForwardPipeline,
     ForwardFlow as NrForwardFlow,
@@ -39,10 +43,10 @@ import {
     BloomStage as NrBloomStage,
     ReflectionProbeFlow as NrReflectionProbeFlow,
     ReflectionProbeStage as NrReflectionProbeStage,
+    RenderPipeline as NrRenderPipeline,
+    RenderFlow as NrRenderFlow,
+    RenderStage as NrRenderStage,
 } from './index';
-import {
-    RenderQueueDesc
-} from '../index.jsb';
 import * as decors from '../../native-binding/decorators';
 import { ccclass, serializable, editable, type } from '../../core/data/class-decorator';
 import { RenderTexture } from '../../asset/assets/render-texture';
@@ -78,6 +82,16 @@ export const ReflectionProbeFlow: typeof NrReflectionProbeFlow = nr.ReflectionPr
 export type ReflectionProbeFlow = NrReflectionProbeFlow;
 export const ReflectionProbeStage: typeof NrReflectionProbeStage = nr.ReflectionProbeStage;
 export type ReflectionProbeStage = NrReflectionProbeStage;
+export const RenderPipeline: typeof NrRenderPipeline = nr.RenderPipeline;
+export type RenderPipeline = NrRenderPipeline;
+export const RenderFlow: typeof NrRenderFlow = nr.RenderFlow;
+export type RenderFlow = NrRenderFlow;
+export const RenderStage: typeof NrRenderStage = nr.RenderStage;
+export type RenderStage = NrRenderStage;
+
+legacyCC.RenderFlow = RenderFlow;
+legacyCC.RenderStage = RenderStage;
+legacyCC.RenderPipeline = RenderPipeline;
 
 interface IRenderPipelineInfo {
     flows: any[];
@@ -348,6 +362,82 @@ class RenderTextureConfig {
     public texture: RenderTexture | null = null;
 }
 
+export enum RenderQueueSortMode {
+    FRONT_TO_BACK,
+    BACK_TO_FRONT,
+}
+ccenum(RenderQueueSortMode);
+
+export class RenderQueueDesc {
+
+    /**
+ * @en Whether the render queue is a transparent queue
+ * @zh 当前队列是否是半透明队列
+    */
+    @serializable
+    @editable
+    public isTransparent = false;
+
+    /**
+     * @en The sort mode of the render queue
+     * @zh 渲染队列的排序模式
+     */
+    @type(RenderQueueSortMode)
+    public sortMode: RenderQueueSortMode = RenderQueueSortMode.FRONT_TO_BACK;
+
+    /**
+ * @en The stages using this queue
+ * @zh 使用当前渲染队列的阶段列表
+ */
+    @type([CCString])
+    public stages: string[] = [];
+
+    constructor() {
+        this.stages = [];
+    }
+
+    public init(): any {
+        return new nr.RenderQueueDesc(this.isTransparent, this.sortMode, this.stages);
+    }
+}
+
+function proxyArrayAttributeImpl(proto: any, attr: string): void {
+    const proxyTarget = `_${attr}_target`;
+    let arrayProxy = (self, targetArrayAttr: string): any => {
+        return new Proxy(self[targetArrayAttr], {
+            get(targetArray, prop, receiver): any {
+                return Reflect.get(targetArray, prop, receiver);
+            },
+            set(targetArray, prop, receiver): boolean {
+                const ret = Reflect.set(targetArray, prop, receiver);
+                self[targetArrayAttr] = targetArray;
+                return ret;
+            }
+        });
+    }
+
+    Object.defineProperty(proto, attr, {
+        configurable: true,
+        enumerable: true,
+        get: function () {
+            this[proxyTarget] ||= [];
+            return arrayProxy(this, proxyTarget);
+        },
+        set: function (v) {
+            this[proxyTarget] = v;
+        }
+    });
+}
+
+
+let proxyArrayAttribute = proxyArrayAttributeImpl;
+
+if (!OPEN_HARMONY) {
+    // WORKAROUND: the proxy array getLength crashed on OH platform
+    proxyArrayAttribute(RenderFlow.prototype, '_stages');
+    proxyArrayAttribute(RenderPipeline.prototype, '_flows');
+}
+
 //-------------------- register types -------------------- 
 
 const Material = jsb.Material;
@@ -367,3 +457,10 @@ decors.patch_ReflectionProbeFlow({ReflectionProbeFlow});
 
 decors.patch_ForwardPipeline({ForwardPipeline, RenderTextureConfig});
 decors.patch_DeferredPipeline({DeferredPipeline, RenderTextureConfig});
+
+decors.patch_RenderQueueDesc({RenderQueueDesc, RenderQueueSortMode, CCString});
+decors.patch_RenderStage({RenderStage});
+
+decors.patch_RenderFlow({RenderFlow, RenderStage});
+
+decors.patch_cc_RenderPipeline({RenderPipeline, RenderFlow});
